+++ ProyectoGit/src/java/clases/Prueba.java
@@ -1,30 +1,9 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
 package clases;
 
-
-import java.util.logging.Level;
-import java.util.logging.Logger;
-import modelo.EventoDAO;
-import modelo.EventoDAOImpl;
-
-/**
- *
- * @author Cristian
- */
 public class Prueba {
     /*
     public static void main(String[] args) {
         
-        EventoDAO dao = new EventoDAOImpl();
-        try {
-            dao.insertarEvento("exs", "es","03-03-2013", "12:50", "13:00", "descripcion", "Auditorio 105","Cesario");
-        } catch (Exception ex) {
-            Logger.getLogger(Prueba.class.getName()).log(Level.SEVERE, null, ex);
-        }
         
         
     }
