+++ ProyectoGit/src/java/dao/DepartamentoDAO.java
@@ -1,4 +1,4 @@
-package modelo;
+package dao;
 
 import java.util.List;
