+++ ProyectoGit/src/java/dao/EventoDAO.java
@@ -1,4 +1,4 @@
-package modelo;
+package dao;
 
 import java.sql.Date;
 import java.sql.ResultSet;
