+++ ProyectoGit/src/java/impl/EventoDAOImpl.java
@@ -1,5 +1,6 @@
-package modelo;
+package impl;
 
+import dao.EventoDAO;
 import controlador.Conexion;
 import java.sql.Date;
 import java.sql.PreparedStatement;
