+++ ProyectoGit/src/java/impl/LugarDAOImpl.java
@@ -1,5 +1,6 @@
-package modelo;
+package impl;
 
+import dao.LugarDAO;
 import controlador.Conexion;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
